+++ src/client/adapters/comunica/mod.ts
@@ -1,2 +1,1 @@
 export * from "./comunica-sparql-engine.ts";
-export * from "./create-comunica-sparql-engine-factory.ts";