+++ src/client/adapters/denokv/mod.ts
@@ -1,3 +1,4 @@
 export * from "./denokv-quad-store.ts";
 export * from "./create-denokv-adapter.ts";
 export * from "./denokv-search-index.ts";
+export * from "./denokv-rdfjs-store.ts";